Define agents and agent clusters
Define the infrastructure for SharedArrayBuffer. This also clarifies
along which boundaries a browser implementation can use processes and
threads.

shaky foundation. Because of that, similar-origin window agents are
the best place to store state that would formerly go on unit of
related similar-origin browsing contexts.

https://github.com/tc39/ecma262/issues/882 is follow-up to define
agents in more detail; in particular make their implicit realms slot
explicit. https://github.com/w3c/css-houdini-drafts/issues/224 is
follow-up to define worklet ownership better which is needed to
define how they relate to agent (sub)clusters.

Fixes part of #2260. Fixes
https://github.com/w3c/ServiceWorker/issues/1115. Fixes most of
https://github.com/w3c/css-houdini-drafts/issues/380 (no tests and no
nice grouping of multiple realms in a single agent as that is not
needed).